refactor: [#4] 프로필 완성 페이지 디자인 변경에 따라 수정
